requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.0.6
django-allauth==0.63.3
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-filter==24.2
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-oauth-toolkit==2.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
jwcrypto==1.5.6
Markdown==3.6
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==70.1.0
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
